refactor(browser): move cef handlers out of utils
