+++ CefFlashBrowser/Views/ViewSourceWindow.xaml.cs
@@ -5,6 +5,7 @@ using CefSharp;
 using SimpleMvvm.Messaging;
 using System.ComponentModel;
 using System.Windows;
+using BrowserHandlers = CefFlashBrowser.Browser.Handlers;
 
 namespace CefFlashBrowser.Views
 {
@@ -42,10 +43,10 @@ namespace CefFlashBrowser.Views
         {
             InitializeComponent();
 
-            browser.DragHandler = new Utils.Handlers.DisableDragHandler();
-            browser.MenuHandler = new Utils.Handlers.ContextMenuHandler();
-            browser.JsDialogHandler = new Utils.Handlers.JsDialogHandler();
-            browser.DownloadHandler = new Utils.Handlers.DownloadHandler();
+            browser.DragHandler = new BrowserHandlers.DisableDragHandler();
+            browser.MenuHandler = new BrowserHandlers.ContextMenuHandler();
+            browser.JsDialogHandler = new BrowserHandlers.JsDialogHandler();
+            browser.DownloadHandler = new BrowserHandlers.DownloadHandler();
             browser.LifeSpanHandler = new ViewSourceBrowserLifeSpanHandler();
 
             Messenger.Global.Register(MessageTokens.CLOSE_ALL_BROWSERS, CloseBrowserHandler);
